fix js erros
fix errors
