+++ index.html
@@ -6,7 +6,13 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Afri-tech Summit</title>
     <link rel="stylesheet" href="css/style.css" />
-
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.0/css/all.min.css"
+      integrity="sha512-xh6O/CkQoPOWDdYTDqeRdPCVd1SpvCA9XXcUnZS2FmJNp1coAFzvtCN9BmamE+4aHK8yyUHUSCcJHgXloTyT2A=="
+      crossorigin="anonymous"
+      referrerpolicy="no-referrer"
+    />
   </head>
   <body>
    
@@ -132,6 +138,7 @@
         </div>
         <div class="speakers-cards" id="speakers-cards">
         </div>
+        </div>
     </section>
     <section class="partners-section" id="partners">
       <div class="partners-container">
@@ -155,6 +162,6 @@
         designed by <a href="https://www.behance.net/adagio07"> Cindy Shin</a>
       </p>
     </footer>
-    <script type="module" src="js/index.js"></script>
+    <script src="js/index.js"></script>
   </body>
 </html>
