+++ js/index.js
@@ -1,4 +1,4 @@
-import speakers from './speakersData.js';
+// import speakers from 'speakersData.js';
 
 const nav = document.getElementById('hamburger');
 const popup = document.getElementById('popup');
@@ -7,6 +7,57 @@ nav.addEventListener('click', () => {
   popup.classList.toggle('hide-popup');
   popup.style.left = 0;
 });
+
+const speakers = [
+  {
+    id: '1',
+    name: 'Rabecca Nabwire',
+    role: 'Co-Founder and Lead software Engineer at Afritech R&D center',
+    description:
+      'lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod.',
+    image: 'images/speaker_01.png',
+  },
+  {
+    id: '2',
+    name: 'Rabecca Nabwire',
+    role: 'Co-Founder and Lead software Engineer at Afritech R&D center',
+    description:
+      'lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod.',
+    image: 'images/speaker_02.png',
+  },
+  {
+    id: '3',
+    name: 'Rabecca Nabwire',
+    role: 'Co-Founder and Lead software Engineer at Afritech R&D center',
+    description:
+      'lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod.',
+    image: 'images/speaker_03.png',
+  },
+  {
+    id: '4',
+    name: 'Rabecca Nabwire',
+    role: 'Co-Founder and Lead software Engineer at Afritech R&D center',
+    description:
+      'lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod.',
+    image: 'images/speaker_04.png',
+  },
+  {
+    id: '5',
+    name: 'Rabecca Nabwire',
+    role: 'Co-Founder and Lead software Engineer at Afritech R&D center',
+    description:
+      'lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod.',
+    image: 'images/speaker_05.png',
+  },
+  {
+    id: '6',
+    name: 'Rabecca Nabwire',
+    role: 'Co-Founder and Lead software Engineer at Afritech R&D center',
+    description:
+      'lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod.',
+    image: 'images/speaker_06.png',
+  },
+];
 
 const data = speakers;
 const speakersCards = document.getElementById('speakers-cards');
