--- js/speakersData.js
@@ -1,53 +1,0 @@
-
-const speakers = [
-  {
-    id: '1',
-    name: 'Rabecca Nabwire',
-    role: 'Co-Founder and Lead software Engineer at Afritech R&D center',
-    description:
-      'lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod.',
-    image: 'images/speaker_01.png',
-  },
-  {
-    id: '2',
-    name: 'Rabecca Nabwire',
-    role: 'Co-Founder and Lead software Engineer at Afritech R&D center',
-    description:
-      'lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod.',
-    image: 'images/speaker_02.png',
-  },
-  {
-    id: '3',
-    name: 'Rabecca Nabwire',
-    role: 'Co-Founder and Lead software Engineer at Afritech R&D center',
-    description:
-      'lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod.',
-    image: 'images/speaker_03.png',
-  },
-  {
-    id: '4',
-    name: 'Rabecca Nabwire',
-    role: 'Co-Founder and Lead software Engineer at Afritech R&D center',
-    description:
-      'lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod.',
-    image: 'images/speaker_04.png',
-  },
-  {
-    id: '5',
-    name: 'Rabecca Nabwire',
-    role: 'Co-Founder and Lead software Engineer at Afritech R&D center',
-    description:
-      'lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod.',
-    image: 'images/speaker_05.png',
-  },
-  {
-    id: '6',
-    name: 'Rabecca Nabwire',
-    role: 'Co-Founder and Lead software Engineer at Afritech R&D center',
-    description:
-      'lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod.',
-    image: 'images/speaker_06.png',
-  },
-];
-
-export default speakers;
